requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blessed==1.20.0
click==8.1.7
colorama==0.4.6
crawlerdetect==0.1.7
fastapi==0.115.5
fuzzywuzzy==0.18.0
gitignore-parser==0.1.11
h11==0.14.0
httptools==0.6.4
idna==3.10
imy==0.4.0
introspection==1.9.2
isort==5.13.2
jaraco-classes==3.4.0
keyring==24.3.1
langcodes==3.5.0
language-data==1.3.0
levenshtein==0.26.1
marisa-trie==1.2.1
more-itertools==10.5.0
multidict==6.1.0
narwhals==1.15.1
ordered-set==4.1.0
path-imports==1.1.2
pillow==10.4.0
propcache==0.2.1
pydantic==2.10.2
pydantic-core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-levenshtein==0.26.1
python-multipart==0.0.19
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.10.1
readchar==4.2.1
revel==0.9.1
rio-ui==0.10.8
sentinel==1.0.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
timer-dict==1.0.0
tomlkit==0.13.2
typing-extensions==4.12.2
unicall==0.1.5
uniserde==0.3.14
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
yarl==1.18.3
